perf(inbound): trim reply startup imports (#51988)
* perf(inbound): narrow reply startup imports

* perf(inbound): trim reply startup imports

* fix(perf): dedupe archive helpers

(cherry picked from commit 177136c9643303737fe5021d1dbdc4a672093563)
