+++ src/auto-reply/reply/directive-handling.parse.ts
@@ -4,8 +4,8 @@ import type { MsgContext } from "../templating.js";
 import type { VerboseLevel } from "./directives.js";
 import { extractStatusDirective, extractVerboseDirective } from "./directives.js";
 import { stripMentions, stripStructuralPrefixes } from "./mentions.js";
-import type { QueueDropPolicy, QueueMode } from "./queue.js";
-import { extractQueueDirective } from "./queue.js";
+import { extractQueueDirective } from "./queue/directive.js";
+import type { QueueDropPolicy, QueueMode } from "./queue/types.js";
 
 export type InlineDirectives = {
   cleaned: string;
